# specs/001-administration/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Echoroo Administration API
  description: Authentication, user management, and project management endpoints
  version: 1.0.0
  contact:
    name: Echoroo Team
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /projects:
    get:
      tags: [projects]
      summary: List projects
      description: Get all projects accessible by current user
      operationId: listProjects
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '401':
          description: Not authenticated

    post:
      tags: [projects]
      summary: Create project
      description: Create a new project
      operationId: createProject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Validation error
        '401':
          description: Not authenticated
  /projects/{projectId}:
    get:
      tags: [projects]
      summary: Get project
      description: Get project details
      operationId: getProject
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          description: Not authenticated
        '403':
          description: Access denied
        '404':
          description: Project not found

    patch:
      tags: [projects]
      summary: Update project
      description: Update project settings (admin only)
      operationId: updateProject
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '401':
          description: Not authenticated
        '403':
          description: Not project admin
        '404':
          description: Project not found

    delete:
      tags: [projects]
      summary: Delete project
      description: Delete project (owner only)
      operationId: deleteProject
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Project deleted
        '401':
          description: Not authenticated
        '403':
          description: Not project owner
        '404':
          description: Project not found
  /projects/{projectId}/members:
    get:
      tags: [projects]
      summary: List project members
      description: Get all members of a project
      operationId: listProjectMembers
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectMemberResponse'
        '401':
          description: Not authenticated
        '403':
          description: Access denied
        '404':
          description: Project not found

    post:
      tags: [projects]
      summary: Add project member
      description: Invite user to project (admin only)
      operationId: addProjectMember
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMemberAddRequest'
      responses:
        '201':
          description: Member added or invitation sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberResponse'
        '400':
          description: User already member
        '401':
          description: Not authenticated
        '403':
          description: Not project admin
        '404':
          description: Project not found
  /projects/{projectId}/members/{userId}:
    patch:
      tags: [projects]
      summary: Update member role
      description: Change member role (admin only)
      operationId: updateProjectMemberRole
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMemberUpdateRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberResponse'
        '400':
          description: Cannot change owner role
        '401':
          description: Not authenticated
        '403':
          description: Not project admin
        '404':
          description: Member not found

    delete:
      tags: [projects]
      summary: Remove project member
      description: Remove member from project (admin only)
      operationId: removeProjectMember
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Member removed
        '400':
          description: Cannot remove owner
        '401':
          description: Not authenticated
        '403':
          description: Not project admin
        '404':
          description: Member not found

  # ============================================
  # Admin Endpoints
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProjectCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
        target_taxa:
          type: string
          maxLength: 500
        visibility:
          type: string
          enum: [private, public]
          default: private
    ProjectUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
        target_taxa:
          type: string
          maxLength: 500
        visibility:
          type: string
          enum: [private, public]
    ProjectMemberAddRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, member, viewer]
          default: member
    ProjectMemberUpdateRequest:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [admin, member, viewer]
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        display_name:
          type: string
        organization:
          type: string
        is_active:
          type: boolean
        is_superuser:
          type: boolean
        is_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time
...
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        target_taxa:
          type: string
        visibility:
          type: string
          enum: [private, public]
        owner:
          $ref: '#/components/schemas/UserResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProjectMemberResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user:
          $ref: '#/components/schemas/UserResponse'
        role:
          type: string
          enum: [admin, member, viewer]
        joined_at:
          type: string
          format: date-time
...